move constants into a separate file; update usages
