+++ dashboard-ui/src/components/SignIn.js
@@ -13,7 +13,7 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 
 import Alert from './Alert';
-import { formSubmission } from '../utils';
+import { formSubmission } from '../constants';
 
 const useStyles = makeStyles(theme => ({
     paper: {
