+++ dashboard-ui/src/hooks/useFormWithSteps.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { formSubmission } from '../utils';
+import { formSubmission } from '../constants';
 import useForm from './useForm';
 
 const useFormWithSteps = (initialValues, callback, entity) => {
