+++ dashboard-ui/src/hooks/useSquadHubData.js
@@ -3,7 +3,7 @@ import { useQuery } from 'react-query';
 import { fetchSquadHubData } from '../clients/ClubClient';
 import { useUserAuth } from '../context/authProvider';
 import { useCurrentClub } from '../context/clubProvider';
-import { queryKeys } from '../utils';
+import { queryKeys } from '../constants';
 
 export default function() {
     const { authData } = useUserAuth();
