+++ dashboard-ui/src/stories/utils/storyDataGenerators.js
@@ -1,7 +1,7 @@
 import { faker } from '@faker-js/faker';
 import _ from 'lodash';
 
-import { playerAttributes } from '../../utils';
+import { playerAttributes } from '../../constants';
 
 // TODO: 31/01/22 remove all usage of these constants from stories and then remove the export keyword
 export const MAX_ATTR_VALUE = 20;
